+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 0.25.0 - RegEx name validation
 
 - Add RegEx name validation
-- Allow to retrieve the plugin version from LibreLoginProvider
+- Allow to retrieve the plugin version from LibreLoginNextProvider
+- Rebranded everything to LibreLoginNext